Added $ to mask Root...
If you are not running on a RPI och a ATB the "grep /" will not give the correct values.
If you add the "$" to the grep it will only pull the line that "ends" with a "/" and that's usually the Root disk. :)
Works for me on my NUC at least. :)